tests/wpt/tests/css/css-masonry/tentative/parsing/masonry-track-valid.html

<!DOCTYPE html>
<title>CSS Masonry: masonry-track-* parsing</title>
<link rel="author" title="Sam Davis Omekara Jr." href="mailto:samomekarajr@microsoft.com">
<link rel="help" href="https://drafts.csswg.org/css-grid-3">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script src="/css/support/shorthand-testcommon.js"></script>
<body>
<script>
  // auto
  test_valid_value("masonry-track-end", "AuTo", "auto");
  test_valid_value("masonry-track-start", "AuTo", "auto");

  // <custom-ident>
  test_valid_value("masonry-track-end", "a_3");
  test_valid_value("masonry-track-start", "-_π");

  // <integer> && <custom-ident>?
  test_valid_value("masonry-track-end", "Z -44", "-44 Z");
  test_valid_value("masonry-track-start", "+90", "90");

  // span && [ <integer> || <custom-ident> ]
  test_valid_value("masonry-track-end", "SpAn i 2", "span 2 i");
  test_valid_value("masonry-track-start", "2 i span", "span 2 i");

  // masonry-track shorthand
  test_shorthand_value("masonry-track", "auto / auto", {
    "masonry-track-start": "auto",
    "masonry-track-end": "auto"
  });
  test_shorthand_value("masonry-track", "auto", {
      "masonry-track-start": "auto",
      "masonry-track-end": "auto"
  });

  test_shorthand_value("masonry-track", "10 / auto", {
      "masonry-track-start": "10",
      "masonry-track-end": "auto"
  });
  test_shorthand_value("masonry-track", "10", {
      "masonry-track-start": "10",
      "masonry-track-end": "auto"
  });

  test_shorthand_value("masonry-track", "-10 / auto", {
      "masonry-track-start": "-10",
      "masonry-track-end": "auto"
  });
  test_shorthand_value("masonry-track", "-10", {
      "masonry-track-start": "-10",
      "masonry-track-end": "auto"
  });

  test_shorthand_value("masonry-track", "span 2 / auto", {
      "masonry-track-start": "span 2",
      "masonry-track-end": "auto"
  });
  test_shorthand_value("masonry-track", "span 2", {
      "masonry-track-start": "span 2",
      "masonry-track-end": "auto"
  });

  test_shorthand_value("masonry-track", "3 last / auto", {
      "masonry-track-start": "3 last",
      "masonry-track-end": "auto"
  });
  test_shorthand_value("masonry-track", "3 last", {
      "masonry-track-start": "3 last",
      "masonry-track-end": "auto"
  });

  test_shorthand_value("masonry-track", "span first / auto", {
      "masonry-track-start": "span first",
      "masonry-track-end": "auto"
  });
  test_shorthand_value("masonry-track", "span first", {
      "masonry-track-start": "span first",
      "masonry-track-end": "auto"
  });
  test_shorthand_value("masonry-track", "span 2 first / auto", {
      "masonry-track-start": "span 2 first",
      "masonry-track-end": "auto"
  });
  test_shorthand_value("masonry-track", "span 2 first", {
      "masonry-track-start": "span 2 first",
      "masonry-track-end": "auto"
  });

  test_shorthand_value("masonry-track", "last / last", {
      "masonry-track-start": "last",
      "masonry-track-end": "last"
  });
  test_shorthand_value("masonry-track", "last", {
      "masonry-track-start": "last",
      "masonry-track-end": "last"
  });
</script>
